Move class phase change in its own file
